+++ packages/deployer/internal/generate-contracts.js
@@ -1,0 +1,20 @@
+const fs = require('fs');
+const path = require('path');
+const glob = require('glob');
+const Mustache = require('mustache');
+
+function getGeneratedContractPaths(config) {
+  return glob
+    .sync(path.join(config.paths.sources, 'Gen*.sol'))
+    .filter((file) => startsWithWord(path.basename(file, '.sol'), 'Gen'));
+}
+
+function renderTemplate(filepath, data = {}) {
+  const template = fs.readFileSync(filepath).toString();
+  return Mustache.render(template, data);
+};
+
+module.exports = {
+  renderTemplate,
+  getGeneratedContractPaths,
+};
