+++ packages/deployer/subtasks/upgrade-proxy.js
@@ -3,7 +3,7 @@ const prompter = require('@synthetixio/core-js/utils/prompter');
 const path = require('path');
 const relativePath = require('@synthetixio/core-js//utils/relative-path');
 const { subtask } = require('hardhat/config');
-const { processTransaction, processReceipt } = require('../utils/transactions');
+const { processTransaction, processReceipt } = require('../internal/process-transactions');
 const { getRouterName } = require('../utils/router');
 const { getProxyPath } = require('../utils/deployments');
 const { SUBTASK_UPGRADE_PROXY, SUBTASK_DEPLOY_CONTRACT } = require('../task-names');
