+++ packages/deployer/utils/deployments.js
@@ -5,43 +5,38 @@ const naturalCompare = require('string-natural-compare');
 const relativePath = require('@synthetixio/core-js/utils/relative-path');
 const { getRouterName } = require('./router');
 const { startsWithWord } = require('@synthetixio/core-js/utils/string');
+const { defaults } = require('../extensions/config');
 
 // Regex for deployment file formats, e.g.: 2021-08-31-00-sirius.json
 const DEPLOYMENT_FILE_FORMAT = /^[0-9]{4}-[0-9]{2}-[0-9]{2}-[0-9]{2,}(?:-[a-z0-9]+)?\.json$/;
 
-function getGeneratedContractPaths(config) {
-  return glob
-    .sync(path.join(config.paths.sources, 'Gen*.sol'))
-    .filter((file) => startsWithWord(path.basename(file, '.sol'), 'Gen'));
-}
-
-function getMainProxyAddress(config, { network = 'local', instance = 'official' } = {}) {
-  const paths = getDeploymentPaths(config, { network, instance });
+function getMainProxyAddress({ network = 'local', instance = 'official' } = {}) {
+  const paths = getDeploymentPaths({ network, instance });
   const data = getDeploymentData(paths.deployments);
   return data.contracts[getProxyPath(config)].deployedAddress;
 }
 
-function getDeploymentData(deploymentsFolder) {
-  const file = getDeploymentFile(deploymentsFolder);
+// function getDeploymentData(deploymentsFolder) {
+//   const file = getDeploymentFile(deploymentsFolder);
 
-  if (!file) {
-    throw new Error(`No deployment file found on "${deploymentsFolder}".`);
-  }
+//   if (!file) {
+//     throw new Error(`No deployment file found on "${deploymentsFolder}".`);
+//   }
 
-  return JSON.parse(fs.readFileSync(file, 'utf8'));
-}
+//   return JSON.parse(fs.readFileSync(file, 'utf8'));
+// }
 
-function getDeploymentFile(deploymentsFolder) {
-  const deployments = getDeploymentFiles(deploymentsFolder);
-  return deployments.length > 0 ? deployments[deployments.length - 1] : null;
-}
+// function getDeploymentFile(deploymentsFolder) {
+//   const deployments = getDeploymentFiles(deploymentsFolder);
+//   return deployments.length > 0 ? deployments[deployments.length - 1] : null;
+// }
 
-function getDeploymentFiles(deploymentsFolder) {
-  return glob
-    .sync(`${deploymentsFolder}/*.json`)
-    .filter((file) => DEPLOYMENT_FILE_FORMAT.test(path.basename(file)))
-    .sort(naturalCompare);
-}
+// function getDeploymentFiles(deploymentsFolder) {
+//   return glob
+//     .sync(`${deploymentsFolder}/*.json`)
+//     .filter((file) => DEPLOYMENT_FILE_FORMAT.test(path.basename(file)))
+//     .sort(naturalCompare);
+// }
 
 function getModulesPaths(config) {
   return glob
@@ -56,14 +51,61 @@ function getProxyPath(config) {
   );
 }
 
-function getDeploymentPaths(config, { network = 'local', instance = 'official' } = {}) {
-  return {
-    deployments: path.join(config.deployer.paths.deployments, network, instance),
-    router: relativePath(
-      path.join(config.paths.sources, `${getRouterName({ network, instance })}.sol`),
-      config.paths.root
-    ),
-  };
+// function getDeploymentPaths({
+//   network = 'local',
+//   instance = 'official',
+//   folder = defaults.paths.deployments,
+//   sources = 'contracts',
+// } = {}) {
+//   return {
+//     deployments: path.join(folder, network, instance),
+//     router: path.join(sources, `${getRouterName({ network, instance })}.sol`),
+//   };
+// }
+
+function getCurrentDeploymentDataForInstance({
+  network = 'local',
+  instance = 'official',
+  deploymentsFolder = defaults.paths.deployments,
+}) {
+  const file = getCurrentDeploymentFileForInstance(deploymentsFolder);
+
+  if (!file) {
+    throw new Error(`No deployment file found on "${deploymentsFolder}".`);
+  }
+
+  return JSON.parse(fs.readFileSync(file, 'utf8'));
+}
+
+function getCurrentDeploymentFileForInstance({
+  network = 'local',
+  instance = 'official',
+  deploymentsFolder = defaults.paths.deployments,
+}) {
+  const deployments = getAllDeploymentsFilesForInstance(deploymentsFolder);
+
+  return deployments.length > 0 ? deployments[deployments.length - 1] : null;
+}
+
+function getAllDeploymentsFilesForInstance(
+  network = 'local',
+  instance = 'official',
+  deploymentsFolder = defaults.paths.deployments,
+) {
+  const instanceFolder = getDeploymentFolderForInstance({ network, instance, deploymentsFolder });
+
+  return glob
+    .sync(`${deploymentsFolder}/*.json`)
+    .filter((file) => DEPLOYMENT_FILE_FORMAT.test(path.basename(file)))
+    .sort(naturalCompare);
+}
+
+function getDeploymentFolderForInstance({
+  network = 'local',
+  instance = 'official',
+  deploymentsFolder = defaults.paths.deployments,
+}) {
+  return path.join(deploymentsFolder, network, instance);
 }
 
 module.exports = {
@@ -74,5 +116,4 @@ module.exports = {
   getDeploymentPaths,
   getModulesPaths,
   getProxyPath,
-  getGeneratedContractPaths,
 };
